added birthdate to finalize onboarding

+++ src/main/kotlin/com/hedvig/underwriter/model/Quote.kt
@@ -65,10 +65,10 @@ fun String.birthDateStringFromNorwegianSsn(): String {
     val trimmedInput = this.trim().replace("-", "").replace(" ", "")
     val day = trimmedInput.substring(0, 2)
     val month = trimmedInput.substring(2, 4)
-    val twoDigitYear = trimmedInput.substring(4, 6).toInt()
+    val twoDigitYear = trimmedInput.substring(4, 6)
     val breakPoint = LocalDate.now().minusYears(10).year.toString().substring(2, 4).toInt()
 
-    val year = if (twoDigitYear > breakPoint) {
+    val year = if (twoDigitYear.toInt() > breakPoint) {
         "19$twoDigitYear"
     } else {
         "20$twoDigitYear"
